Create calculator financial
I created a simple calculator for investors. I imported the yfinance library, and then researched exchange rates and set fixed exchange rates. In this project, I wanted the person to choose any type of stock (it may not be the stocks shown in the examples), and the current value of the stock is displayed, and then it asks how much we invested and then what the currency was. If the currency was not the dollar, then we would use our exchange rates to convert the money to dollars. Then, the value in dollars is shown, and then how many shares we purchased, taking into account that they can be fractional shares. If you want to test with more shares, you can go to the yfinance website and search for more shares, remembering that you have to enter the abbreviation and not the name of the company.